elf2bestape: Do not flattern relative relocation by default
GCC may emit these in some cases where it needs to emit a long jump, and so far vast majority (as in all I've seen so far) of these type of relocations have been handled by the generated code instead of the loader.

The old behavior can be enabled by passing the flag --force-flattern-rel, although this may also be eventually removed in the future.