elf2bestape: Change how timestamp values are produced in deterministic mode
Deterministic mode now actually measures the built PE file and generates a hash which would then be used to fill the timestamp field. This is conceptually similar to what later versions of MSVC would do but the algorithm is different.

Note that the measurement algorithm is still not final and subjects to change in the future.